+++ src/handlers/wallets/multisig.rs
@@ -12,12 +12,12 @@ pub fn index(state: &ServerState) -> JsonResult {
 }
 
 #[get("/multisig_wallets/<id>/get_utxos?<params>")]
-pub fn get_utxos(state: &ServerState, id: u64, params: GetTransactionParams) -> JsonResult {
+pub fn get_utxos(state: &ServerState, id: usize, params: GetTransactionParams) -> JsonResult {
     MultisigWallet::get_utxos(state, id, params.limit, params.since)
 }
 
 #[get("/multisig_wallets/<id>/get_incoming?<params>")]
-pub fn get_incoming(state: &ServerState, id: u64, params: GetTransactionParams) -> JsonResult {
+pub fn get_incoming(state: &ServerState, id: usize, params: GetTransactionParams) -> JsonResult {
     MultisigWallet::get_incoming(state, id, params.limit, params.since)
 }
